+++ Extension/src/background/services/ui/popup/popup-common.ts
@@ -43,8 +43,6 @@ import {
 } from '../../../api';
 import { tabsApi as tsWebExtTabsApi } from '../../../tswebextension';
 import { UserAgent } from '../../../../common/user-agent';
-import { ABTestManager } from '../../telemetry';
-import { AG_51010_LIMITATIONS_BROWSER_B } from '../../telemetry/abtest/constants';
 import { SearchPageAccessService } from '../../searchPageAccessService';
 
 /**
@@ -119,11 +117,6 @@ export type GetTabInfoForPopupResponse = {
         hasUserRulesToReset: boolean;
 
         /**
-         * State for AG-51010 A/B test «Limitations of browser blocking».
-         */
-        showAlternativeProtectionButton: boolean;
-
-        /**
          * Whether to show Opera search access warning.
          */
         showSearchAccessWarning: boolean;
@@ -189,7 +182,6 @@ export abstract class PopupServiceCommon {
                 notification: await promoNotificationApi.getCurrentNotification(),
                 isDisableShowAdguardPromoInfo: settingsStorage.get(SettingOption.DisableShowAdguardPromoInfo),
                 hasUserRulesToReset: await UserRulesApi.hasRulesForUrl(tabContext.info.url),
-                showAlternativeProtectionButton: await ABTestManager.hasVariant(AG_51010_LIMITATIONS_BROWSER_B),
                 showSearchAccessWarning: await SearchPageAccessService.shouldShowNotification(),
             },
         };