+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `$urltransform` modifier support [tsurlfilter#111].
+- `$removeparam` support for SPA navigations [tsurlfilter#188].
 
 ### Fixed
 
@@ -31,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#2701]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/2701
 [#3533]: https://github.com/AdguardTeam/AdguardBrowserExtension/issues/3533
 [tsurlfilter#111]: https://github.com/AdguardTeam/tsurlfilter/issues/111
+[tsurlfilter#188]: https://github.com/AdguardTeam/tsurlfilter/issues/188
 [tsurlfilter#189]: https://github.com/AdguardTeam/tsurlfilter/issues/189
 
 ## [5.4 patch 2] - 2026-05-14